requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9

